Pipeline2/requirements.txt
transformers==4.6.1
psycopg2-binary
